michaelr - GOOS, p. 179b; main reimplemented to join multiple auctions. I had to do a lot of trouble shooting to figure out what was going on because my code behaved differently that the book's. ("I don't see this error.", p. 178a)
- I used "triangulation" during my trouble shooting: made 2 tests - 1 for item-54321, another for item-65432.

git-svn-id: svn://localhost/code_practice/trunk@186 50d94a75-6656-bf47-8487-448895b3180b

+++ code_practice_junit/src/main/java/auctionsniper/SwingThreadSniperListener.java
@@ -32,10 +32,11 @@ public class SwingThreadSniperListener implements SniperListener {
   }
 
  	/*>>>REMOVE-BEGIN<<<*/
- 	private int forThisLong = 0;
-	private void sleep(int sleepDuration) {
+ 	//private double forThisLong = 3.5;
+ 	private double forThisLong = 0.0;
+	private void sleep(double sleepDuration) {
 		try {
-			Thread.sleep(sleepDuration * 1000);
+			Thread.sleep((int)(sleepDuration * 1000));
 		} catch(InterruptedException ex) {
 			Thread.currentThread().interrupt();
 		}
